contract/api/v1/serializers.py
from rest_framework import serializers

from contract.models import Contract


class ContractSerializer(serializers.ModelSerializer):
    class Meta:
        model = Contract
        fields = ('id', 'name',)
